fix: address review feedback for Load3D E2E tests
