+++ browser_tests/tests/load3d/Load3DHelper.ts
@@ -45,27 +45,6 @@ export class Load3DHelper {
     }, hex)
   }
 
-  async waitForWidgetValue(
-    nodeId: number,
-    widgetName: string,
-    expected: string
-  ): Promise<void> {
-    await expect
-      .poll(
-        () =>
-          this.node.page().evaluate(
-            ({ nodeId, widgetName }) => {
-              const n = window.app!.graph.getNodeById(nodeId)
-              const w = n?.widgets?.find((w) => w.name === widgetName)
-              return w?.value
-            },
-            { nodeId, widgetName }
-          ),
-        { timeout: 15000 }
-      )
-      .toContain(expected)
-  }
-
   async waitForModelLoaded(): Promise<void> {
     await expect(this.node.getByTestId(TestIds.loading.overlay)).toBeHidden({
       timeout: 30000
