+++ browser_tests/tests/load3d/load3dViewer.spec.ts
@@ -1,24 +1,10 @@
 import { expect } from '@playwright/test'
 
-import { comfyPageFixture as test } from '@e2e/fixtures/ComfyPage'
 import { assetPath } from '@e2e/fixtures/utils/paths'
-import { Load3DHelper } from './Load3DHelper'
-import { Load3DViewerHelper } from './Load3DViewerHelper'
+import { load3dViewerTest as test } from '@e2e/fixtures/helpers/Load3DFixtures'
 
 test.describe('Load3D Viewer', () => {
-  let load3d: Load3DHelper
-  let viewer: Load3DViewerHelper
-
-  test.beforeEach(async ({ comfyPage }) => {
-    await comfyPage.settings.setSetting('Comfy.VueNodes.Enabled', true)
-    await comfyPage.settings.setSetting('Comfy.Load3D.3DViewerEnable', true)
-    await comfyPage.workflow.loadWorkflow('3d/load3d_node')
-    await comfyPage.vueNodes.waitForNodes()
-
-    const node = comfyPage.vueNodes.getNodeLocator('1')
-    load3d = new Load3DHelper(node)
-    viewer = new Load3DViewerHelper(comfyPage.page)
-
+  test.beforeEach(async ({ comfyPage, load3d }) => {
     // Upload cube.obj so the node has a model loaded
     const uploadResponsePromise = comfyPage.page.waitForResponse(
       (resp) => resp.url().includes('/upload/') && resp.status() === 200,
@@ -30,14 +16,19 @@ test.describe('Load3D Viewer', () => {
     await fileChooser.setFiles(assetPath('cube.obj'))
     await uploadResponsePromise
 
-    await load3d.waitForWidgetValue(1, 'model_file', 'cube.obj')
+    const nodeRef = await comfyPage.nodeOps.getNodeRefById(1)
+    const modelFileWidget = await nodeRef.getWidget(0)
+    await expect
+      .poll(() => modelFileWidget.getValue(), { timeout: 5000 })
+      .toContain('cube.obj')
+
     await load3d.waitForModelLoaded()
   })
 
   test(
     'Opens viewer dialog with canvas and controls sidebar',
     { tag: '@smoke' },
-    async () => {
+    async ({ load3d, viewer }) => {
       await load3d.openViewerButton.click()
       await viewer.waitForOpen()
 
@@ -54,7 +45,7 @@ test.describe('Load3D Viewer', () => {
   test(
     'Cancel button closes the viewer dialog',
     { tag: '@smoke' },
-    async () => {
+    async ({ load3d, viewer }) => {
       await load3d.openViewerButton.click()
       await viewer.waitForOpen()
